+++ Sources/OpenKey/win32/OpenKey/OpenKey/SpecialAppsDialogSciter.cpp
@@ -418,7 +418,7 @@ void SpecialAppsDialogSciter::saveData() {
     }
     
     // Notify main process
-    HWND mainWnd = FindWindow(_T("OpenKeyVietnameseInputMethod"), NULL);
+    HWND mainWnd = FindWindow(APP_CLASS, NULL);
     if (mainWnd) {
         PostMessage(mainWnd, WM_USER + 101, 0, 0);
     }